added authentification to check services

+++ src/services/check_service.py
@@ -21,8 +21,11 @@ async def perform_dns_check(monitor_id, dns_resolver, url):
 
 class CheckService():
     @staticmethod
-    async def perform_check(monitor_id):
+    async def perform_check(monitor_id, user_id):
         monitor = await orm.select_monitor(monitor_id)
+        if monitor.user_id != user_id:
+            return {"success": False,
+                    "status_code": 404,}
         if monitor.type_of_request == dns_or_https.https:
             await perform_https_check(monitor.id, monitor.url)
         else:
@@ -35,13 +38,26 @@ class CheckService():
             ]
             for domain_name in domain_list:
                 await perform_dns_check(monitor.id, monitor.url, domain_name)
+        return {"success": True,
+                "status_code": 200}
 
 
     @staticmethod
-    async def get_logs(monitor_id, start, end):
+    async def get_logs(monitor_id, start, end, user_id):
+        monitor = await orm.select_monitor(monitor_id)
+        if monitor.user_id != user_id:
+            return {"success": False,
+                    "status_code": 404}
         logs = await orm.select_check_logs(monitor_id, start, end)
-        return logs
+        return {"success": True,
+                "logs": logs}
 
     @staticmethod
-    async def delete_logs(monitor_id, start, end):
+    async def delete_logs(monitor_id, start, end, user_id):
+        monitor = await orm.select_monitor(monitor_id)
+        if monitor.user_id != user_id:
+            return {"success": False,
+                    "status_code": 404}
         await orm.delete_logs(monitor_id, start, end)
+        return {"success": True,
+                "status_code": 200}
